fixed the bug with pruning %
vox_grid.n_nonempty_voxels was uint_32t when it should be uint_64t

+++ partition/ply_c/ply_c.cpp
@@ -88,7 +88,7 @@ class AttributeGrid {
        index(0)
     {}
     //---methods for the occurence grid---
-    uint32_t n_nonempty_voxels()
+    uint64_t n_nonempty_voxels()
     {
         return this->index;
     }
